fix(types): change REAL from f32 to f64 for SQLite compatibility
SQLite REAL is an 8-byte IEEE float (same as f64), not 4-byte (f32).
This change fixes precision issues in TCL tests by:
- Changing SqlValue::Real from f32 to f64
- Updating binary serialization to read/write 8 bytes for REAL
- Fixing all pattern matching to separate Float (f32) from Real (f64)
- Grouping Real with Double/Numeric (all f64) in combined patterns
- Displaying infinity as 9.0e+999 instead of inf for SQLite compat

+++ crates/vibesql-executor/src/evaluator/window/sorting.rs
@@ -79,12 +79,12 @@ pub fn compare_values(a: &SqlValue, b: &SqlValue) -> Ordering {
         (SqlValue::Character(a), SqlValue::Character(b)) => a.cmp(b),
         (SqlValue::Boolean(a), SqlValue::Boolean(b)) => a.cmp(b),
 
-        // Type coercion for mixed integer/real (Real is f32)
+        // Type coercion for mixed integer/real (Real is now f64)
         (SqlValue::Integer(a), SqlValue::Real(b)) => {
-            (*a as f32).partial_cmp(b).unwrap_or(Ordering::Equal)
+            (*a as f64).partial_cmp(b).unwrap_or(Ordering::Equal)
         }
         (SqlValue::Real(a), SqlValue::Integer(b)) => {
-            a.partial_cmp(&(*b as f32)).unwrap_or(Ordering::Equal)
+            a.partial_cmp(&(*b as f64)).unwrap_or(Ordering::Equal)
         }
 
         // Other type combinations: compare as strings
